refactor: 前端直接嵌入 DocSpace 完整界面，移除自定义 UI
用 iframe 嵌入 DocSpace Portal，文件管理/上传/新建/编辑全部
由 DocSpace 原生 UI 处理，前端代码从 392 行减到 23 行。

移除: DocSpaceFrame, DocumentList, UploadButton, use-upload, use-documents
保留: useDocSpaceConfig（获取 portalUrl + enabled 状态）
