+++ frontend/src/features/docspace/hooks/use-docspace-config.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query'
+import { docspaceApi } from '../api'
+
+export function useDocSpaceConfig() {
+  return useQuery({
+    queryKey: ['docspace', 'config'],
+    queryFn: docspaceApi.getConfig,
+    staleTime: 5 * 60 * 1000,
+  })
+}
