+++ frontend/src/features/docspace/index.ts
@@ -1,6 +1,3 @@
-export { DocSpaceFrame } from './components/DocSpaceFrame'
-export { DocumentList } from './components/DocumentList'
-export { UploadButton } from './components/UploadButton'
-export { useDocSpaceConfig, useDocSpaceDocuments } from './hooks/use-documents'
-export { useUploadDocSpace } from './hooks/use-upload'
-export type { DocSpaceConfig, DocSpaceDocument, DocSpaceUploadResult } from './types'
+export { docspaceApi } from './api'
+export { useDocSpaceConfig } from './hooks/use-docspace-config'
+export type { DocSpaceConfig } from './types'
